Add codes

+++ python/args.py
@@ -6,7 +6,7 @@ def main():
     psr.add_argument("arg1", help="positional argument, required")
     psr.add_argument("-a", "--alpha", help="basic format of optional argument")
     psr.add_argument("--default_value", default="Hello")
-    psr.add_argument("--integer", type=int)
+    psr.add_argument("--integer", type=float)
     psr.add_argument("--choice", choices=["hoge", "fuga", "piyo"])
     psr.add_argument("--args", nargs="*")
     args = psr.parse_args()
